delete character by id work done

+++ Services/CharacterService/CharacterService.cs
@@ -56,11 +56,21 @@ namespace DotNet_RPG.Services.CharacterService
 
             try
             {
-                Character character = await _context.Characters.FirstAsync(c => c.Id == id);
-                _context.Characters.Remove(character);
+                Character character = await _context.Characters
+                    .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
 
-                await _context.SaveChangesAsync();
-                serviceResponse.Data = (_context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+
+                if (character !=null)
+                {
+                     _context.Characters.Remove(character);
+                     await _context.SaveChangesAsync();
+                     serviceResponse.Data = (_context.Characters.Where(c=>c.User.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+                }
+                else
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Character not found....";
+                }
             }
             catch (Exception ex)
             {
